LayoutTests/fast/animation/request-animation-frame-throttle-subframe-display-none-expected.txt

Tests that requestAnimationFrame is throttled in subframes that are display:none

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Frame is initially visibile so requestAnimationFrame should not be throttled
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() became false
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() is false
Hiding subframe.
testFrame.style.display = 'none';
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() became true
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() is true
Show subframe again
testFrame.style.display = 'block';
PASS testFrame.contentWindow.internals.isRequestAnimationFrameThrottled() became false
PASS successfullyParsed is true

TEST COMPLETE
